Add files via upload
I believe I have finished the API. I could add extended features like preparing all the statements, but that would take me a very long time to finish, so this API should do everything we need it to do. It is looking at the creds file from the databaseUtil folder for credentials (You can change this in index.php). 

There are major changes to the putRecord and delRecord functions. They now use String Arrays for the id_name and id_value fields.
There are new restrictions on how we use the API to make sure that malevolent users don't delete everything from our tables and such. Check the bottom of js/ConferenceAPIJs/databaseFunctions.js for a detailed listing of the restrictions on the API.

If there are any problems with this API, let me know. It may be a change that you need to make somewhere down the line or a problem on my end.