Add files via upload
I've added in the new databaseFunctions.js. To use this new js file, you will need to move it into the proper directory (conferenceAPIJs) and also move the corresponding new API php files in conferenceAPI4-30-2019 to conferenceAPI. I've listed the changes to all files in the Changes.txt file included here. I'd like to sit down with everyone to further explain usage for the improved methods, so no need to rush implementing this.